requirements.txt
pandas
numpy
jupyter
faiss-cpu
langchain-experimental
pymongo
PyPDF2
selenium
youtube-transcript-api
langchain-huggingface
bson
PyMUPDF